split refactor-transform into own module
